% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_ActualVapourPressure_for_hourly}
\alias{cal_ActualVapourPressure_for_hourly}
\title{Calculating actual vapour pressure for hourly time step}
\usage{
cal_ActualVapourPressure_for_hourly(Thr, RHhr)
}
\arguments{
\item{Thr}{is average hourly temperature (degrees Celsius).}

\item{RHhr}{is average hourly relative humidity [\%].}
}
\value{
A vector for average hourly actual vapour pressure [kPa].
}
\description{
Calculating actual vapour pressure for hourly time step
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
