% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_ActualVapourPressure_from_dewPoint}
\alias{cal_ActualVapourPressure_from_dewPoint}
\title{Actual vapour pressure derived from dewpoint temperature}
\usage{
cal_ActualVapourPressure_from_dewPoint(Tdew)
}
\arguments{
\item{Tdew}{dew point temperature(degrees Celsius).}
}
\value{
A vector for actual vapour pressure
}
\description{
As the dewpint temperature is the temperature to which the air
   needs to be cooled to make the air saturated, the actual vapour pressure
   is the saturation vapour pressure at the dewpoint temperature.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
