% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_ActualVapourPressure_from_psychrometricData}
\alias{cal_ActualVapourPressure_from_psychrometricData}
\title{Actual vapour pressure (ea) derived from psychrometric data}
\usage{
cal_ActualVapourPressure_from_psychrometricData(Twet, Tdry, P, type)
}
\arguments{
\item{Twet, Tdry}{wet bulb depression, with Tdry the dry bulb and Twet the wet bulb temperature (degrees Celsius).}

\item{P}{is the atmospheric pressure (kPa).}

\item{type}{psychrometer type ("Asmann type","natural ventilated","non-ventilated").}
}
\value{
A vector for Actual vapour pressure (ea)
}
\description{
The actual vapour pressure can be determined from the difference
   between the dry and wet bulb temperatures, the so-called wet bulb depression.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
