% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_ActualVapourPressure_from_RHmax}
\alias{cal_ActualVapourPressure_from_RHmax}
\title{Calculating actual vapour pressure derived from RHmax}
\usage{
cal_ActualVapourPressure_from_RHmax(Tmin, RHmax)
}
\arguments{
\item{Tmin}{daily minimum temperature (degrees Celsius).}

\item{RHmax}{maximum relative humidity (\%).}
}
\value{
A vector for actual vapour pressure
}
\description{
When using equipment where errors in estimating RHmin can be large,
 or when RH data integrity are in doubt, then one should use only RHmax.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
