% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_ActualVapourPressure_from_RHmean}
\alias{cal_ActualVapourPressure_from_RHmean}
\title{Calculating actual vapour pressure derived from RHmean}
\usage{
cal_ActualVapourPressure_from_RHmean(RHmean, Tmax, Tmin)
}
\arguments{
\item{RHmean}{mean relative humidity(\%).}

\item{Tmax}{daily maximum temperature (degrees Celsius).}

\item{Tmin}{daily minimum temperature (degrees Celsius).}
}
\value{
A vector for actual vapour pressure
}
\description{
In the absence of RH max and RHmin, it can be used to estimate
   actual vapour pressure.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
