% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_afterRedistribution}
\alias{cal_afterRedistribution}
\title{Calculating the volumetric water content after redistribution}
\usage{
cal_afterRedistribution(THETA_v_sat, alpha, Ksat, deltaT, L, THETA11_1)
}
\arguments{
\item{THETA_v_sat}{soil saturation water content (m3 m-3)}

\item{alpha}{empirical coefficient. 13 for homogenous soil, 13-16 for heterogeneous soil}

\item{Ksat}{saturated hydraulic conductitity}

\item{deltaT}{time step difference (day)}

\item{L}{the thickness(m) of soil layer i}

\item{THETA11_1}{the volumetric water content before redistribution (m3 m-3)}
}
\value{
A value for the volumetric water content after redistribution(m3 m-3)
}
\description{
Calculating the volumetric water content after redistribution
}
