% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_bulkBoundaryLayerResistance}
\alias{cal_bulkBoundaryLayerResistance}
\title{Calculating bulk boundary layer resistance}
\usage{
cal_bulkBoundaryLayerResistance(nu, u_h, w, L)
}
\arguments{
\item{nu}{the wind speed extinction coefficient (taken as 2)}

\item{u_h}{the wind speed at the canopy top (i.e., at plant height h) (m s-1)}

\item{w}{is the mean leaf width (m)}

\item{L}{leaf area index}
}
\value{
A vector for bulk boundary layer resistance (s/m)
}
\description{
Calculating bulk boundary layer resistance
}
