% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_canopyPenetrationProbabilityForNetRadiation}
\alias{cal_canopyPenetrationProbabilityForNetRadiation}
\title{The canopy penetration probability for net radiation}
\usage{
cal_canopyPenetrationProbabilityForNetRadiation(kRn, L)
}
\arguments{
\item{kRn}{the canopy extinction coefficient for net radiation (taken as 0.3)}

\item{L}{the leaf area index (m2 leaf m-2 ground)}
}
\value{
The vector for canopy penetration probability for net radiation
}
\description{
The canopy penetration probability for net radiation
}
