% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_canopyResistance}
\alias{cal_canopyResistance}
\title{Calculating canopy resistance}
\usage{
cal_canopyResistance(a1, a2, It, L, Lmax)
}
\arguments{
\item{a1, a2}{are empirical coefficients, dependent on the crop type.}

\item{It}{the total hourly solar irradiance (W m-2 ground)}

\item{L}{the leaf area index (m2 leaf m-2 ground)}

\item{Lmax}{the maximum total leaf area index (m2 leaf m-2 ground)}
}
\value{
A vector for canopy resistance(s/m)
}
\description{
Calculating canopy resistance
}
