% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_canopyTem}
\alias{cal_canopyTem}
\title{Calculating canopy temperature}
\usage{
cal_canopyTem(Hc, r_c_a, Hs, r_a_a, rho_cp, Tr)
}
\arguments{
\item{Hc}{crop sensible heat fluxes (W m-2)}

\item{r_c_a}{the bulk boundary layer resistance (s m-1)}

\item{Hs}{soil sensible heat fluxes (W m-2)}

\item{r_a_a}{the aerodynamic resistance between the mean canopy flow and reference level (s m-1)}

\item{rho_cp}{the volumetric heat capacity for air (1221.09 J m-3 K-1)}

\item{Tr}{Tr is the air temperature at reference level (Celsius degree). weather station.}
}
\value{
A vector for the canopy (foliage) temperature (Celsius degree)
}
\description{
Calculating canopy temperature
}
