% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_cropRoughnessLength}
\alias{cal_cropRoughnessLength}
\title{Calculating the crop roughness length}
\usage{
cal_cropRoughnessLength(h)
}
\arguments{
\item{h}{the plant height (m)}
}
\value{
A vector for the crop roughness length(m)
}
\description{
Calculating the crop roughness length
}
