% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_daylightHours}
\alias{cal_daylightHours}
\title{Calculating Daylight hours}
\usage{
cal_daylightHours(sunsetHourAngle)
}
\arguments{
\item{sunsetHourAngle}{is the sunset hour angle in radians from cal_sunsetHourAngle().}
}
\value{
A vector for day light Hours
}
\description{
Calculating Daylight hours
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
