% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DualKc.R
\name{cal_Dei_for_DualKc}
\alias{cal_Dei_for_DualKc}
\title{calculating the depletion in the  topsoil layer  at the end of the day}
\usage{
cal_Dei_for_DualKc(Dei_start, P, I, E, Dep, TEW)
}
\arguments{
\item{Dei_start}{Depletion in the topsoil layer}

\item{P}{Precipitation}

\item{I}{Irrigation}

\item{E}{Evaporation on day i, mm}

\item{Dep}{Deep percolation loss from the topsoil layer on day i
if soil water content exceeds field capacity, mm}

\item{TEW}{Maximum cumulative depth of evaporation (depletion) from the topsoil layer}
}
\value{
A value for the depletion in the topsoil layer at the end of the day
}
\description{
In fact, it performs water balance in a day
}
