% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Single_Kc.R
\name{cal_DP_for_singleKc}
\alias{cal_DP_for_singleKc}
\title{Calculating deep percolation}
\usage{
cal_DP_for_singleKc(P, I, ETa, Dri_start)
}
\arguments{
\item{P}{Precipitation}

\item{I}{Irrigation}

\item{ETa}{Actual evapotranspiration}

\item{Dri_start}{The depletion of root layer}
}
\value{
A value for deep percolation
}
\description{
Calculating deep percolation
}
