% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DualKc.R
\name{cal_DPe_for_DualKc}
\alias{cal_DPe_for_DualKc}
\title{Deep percolation loss from the topsoil layer}
\usage{
cal_DPe_for_DualKc(P, I, Dei_start, fw)
}
\arguments{
\item{P}{Precipitation}

\item{I}{Irrigation}

\item{Dei_start}{Depletion in the topsoil layer}

\item{fw}{Fraction of soil surface wetted by irrigation, 0.01-1}
}
\value{
A value for deep percolation loss from the topsoil layer
}
\description{
Deep percolation loss from the topsoil layer
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
