% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DualKc.R
\name{cal_DPr_for_DualKc}
\alias{cal_DPr_for_DualKc}
\title{Deep percolation loss from the root layer}
\usage{
cal_DPr_for_DualKc(P, Irrigation, ETa, Dri_start)
}
\arguments{
\item{P}{Precipitation}

\item{Irrigation}{Irrigation}

\item{ETa}{Actual evapotranspiration}

\item{Dri_start}{Depletion in the root layer}
}
\value{
A value for deep percolation loss from the root layer
}
\description{
Deep percolation loss from the root layer
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
