% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_eddyDiffusivity_Canopytop}
\alias{cal_eddyDiffusivity_Canopytop}
\title{Calculating eddy diffusivity at the canopy top}
\usage{
cal_eddyDiffusivity_Canopytop(k = 0.4, u_, h)
}
\arguments{
\item{k}{the von Karman constant (0.4)}

\item{u_}{the friction velocity (m/s)}

\item{h}{the plant height (m)}
}
\value{
A vector for eddy diffusivity at the canopy top(m2/s)
}
\description{
Calculating eddy diffusivity at the canopy top
}
