% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_eddyDiffusivity_heightZ}
\alias{cal_eddyDiffusivity_heightZ}
\title{Calculating eddy diffusivity at height z}
\usage{
cal_eddyDiffusivity_heightZ(Kh, nK, z, h)
}
\arguments{
\item{Kh}{eddy diffusivity at the canopy top(m2/s)}

\item{nK}{the eddy diffusivity extinction coefficient (taken as 2)}

\item{z}{height(m)}

\item{h}{the plant height(m)}
}
\value{
A vector for eddy diffusivity at height z (m2/s)
}
\description{
Calculating eddy diffusivity at height z
}
