% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_frictionVelocity}
\alias{cal_frictionVelocity}
\title{Calculating friction velocity}
\usage{
cal_frictionVelocity(k = 0.4, zr = 2, u_zr, d, z0)
}
\arguments{
\item{k}{the von Karman constant (0.4)}

\item{zr}{the height of the weather station(m).}

\item{u_zr}{the wind speed(m/s) at the reference height zr (m).}

\item{d}{zero plane displacement height (m)}

\item{z0}{the crop roughness length(m)}
}
\value{
A vector for friction velocity(m/s)
}
\description{
Calculating friction velocity
}
