% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_hourAngle}
\alias{cal_hourAngle}
\title{Calculating hour angle}
\usage{
cal_hourAngle(th)
}
\arguments{
\item{th}{is the local solar time.}
}
\value{
A vector for hour angle (Radian)
}
\description{
Calculating hour angle
}
\examples{
cal_hourAngle(12)
}
\references{
Teh CBS.Introduction to mathematical modeling of crop growth:
           How the equations are derived and assembled into a computer model.
           Brown Walker Press, 2006.
}
