% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_inverseRelativeDistance_Earth_sun}
\alias{cal_inverseRelativeDistance_Earth_sun}
\title{Calculating inverse relative distance Earth-sun}
\usage{
cal_inverseRelativeDistance_Earth_sun(J)
}
\arguments{
\item{J}{is the number of the day in the year between 1 (1 January) and 365
or 366 (31 December).}
}
\value{
A vector for inverse relative distance Earth-sun (Radian)
}
\description{
Calculating inverse relative distance Earth-sun
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
