% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DualKc.R
\name{cal_Kc_max_for_DualKc}
\alias{cal_Kc_max_for_DualKc}
\title{An upper limit on the evaporation and transpiration from any cropped surface}
\usage{
cal_Kc_max_for_DualKc(u2, RHmin, h, Kcb)
}
\arguments{
\item{u2}{The wind speed at 2 m}

\item{RHmin}{Minimum relative humidity}

\item{h}{Plant height}

\item{Kcb}{Basal crop coefficient}
}
\value{
A vector for the upper limit on the evaporation and transpiration from any cropped surface
}
\description{
It is imposed to reflect the natural constraints placed on
   available energy represented by the energy balance difference Rn - G - H
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
