% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Single_Kc.R
\name{cal_Kcini_for_SingleKc}
\alias{cal_Kcini_for_SingleKc}
\title{Calculating Kcini value}
\usage{
cal_Kcini_for_SingleKc(Pmean, ET0, tw, type, fw)
}
\arguments{
\item{Pmean}{is the average depth od infiltrated water per wetting events(mm)}

\item{ET0}{mean ET0 during initial period(mm/day)}

\item{tw}{is the mean interval between wetting events(days)}

\item{type}{soil type:"coarse soil textures" and "medium and fine soil textures"}

\item{fw}{the fraction of surfaces wetted by irrigation or rain (0-1)}
}
\value{
A value for Kcini value
}
\description{
Calculating Kcini value
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
