% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_localDolarTime}
\alias{cal_localDolarTime}
\title{Calculating local solar time}
\usage{
cal_localDolarTime(td, t, gamma, gamma_sm)
}
\arguments{
\item{td}{The day of year.}

\item{t}{is the local time.}

\item{gamma}{is the local longitude (Radian).}

\item{gamma_sm}{is the standard longitude (Radian).}
}
\value{
A vector for local solar time(Hour)
}
\description{
Local solar time is different with local time.
}
\examples{
cal_localDolarTime(td=1,t=12,gamma=0.52,gamma_sm=2.09)
}
\references{
Teh CBS.Introduction to mathematical modeling of crop growth:
           How the equations are derived and assembled into a computer model.
           Brown Walker Press, 2006.
}
