% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_meanCanopyFlowToReferenceLevel}
\alias{cal_meanCanopyFlowToReferenceLevel}
\title{Calculating mean canopy flow to reference level}
\usage{
cal_meanCanopyFlowToReferenceLevel(k = 0.4, u_, zr, d, h, nK, z0)
}
\arguments{
\item{k}{the von Karman constant (0.4)}

\item{u_}{the friction velocity (m s-1)}

\item{zr}{is the reference height (m).the height of the weather station(m).}

\item{d}{zero plane displacement height (m)}

\item{h}{the plant height(m)}

\item{nK}{the eddy diffusivity extinction coefficient (taken as 2)}

\item{z0}{the crop roughness length (m)}
}
\value{
A vector for mean canopy flow to reference level
}
\description{
Calculating mean canopy flow to reference level
}
