% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_meanSaturationVapourPressure}
\alias{cal_meanSaturationVapourPressure}
\title{Calculating mean saturation vapour pressure}
\usage{
cal_meanSaturationVapourPressure(Tmax, Tmin)
}
\arguments{
\item{Tmax}{the daily maximum  air temperature(degrees Celsius).}

\item{Tmin}{the daily  minimum air temperature(degrees Celsius).}
}
\value{
A vector for mean saturation vapour pressure (es)
}
\description{
Due to the non-linearity of the above equation, the mean saturation
   vapour pressure for a day, week, decade or month should be computed as the
   mean between the saturation vapour pressure at the mean daily maximum and
   minimum air temperatures for that period.
}
\details{
Using mean air temperature instead of daily minimum and maximum
   temperatures results in lower estimates for the mean saturation vapour
   pressure. The corresponding vapour pressure deficit (a parameter expressing
   the evaporating power of the atmosphere) will also be smaller and the result
   will be some underestimation of the reference crop evapotranspiration.
   Therefore,the mean sauration vapour pressure should be calculated as the
   mean between the saturation vapour pressure at both the daily maximum
   and minimum air temperature.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
