% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_netRadiation}
\alias{cal_netRadiation}
\title{Calculating net radiation Rn}
\usage{
cal_netRadiation(Rns, Rnl)
}
\arguments{
\item{Rns}{incoming net shortwave radiation. From cal_netSolarRadiation().}

\item{Rnl}{outgoing net longwave radiation. From cal_netLongwaveRadiation().}
}
\value{
A vector for net radiation
}
\description{
The net radiation (Rn) is the difference between the incoming
   net shortwave radiation (Rns) and the outgoing net longwave radiation (Rnl).
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
