% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_netRadiationForCrop}
\alias{cal_netRadiationForCrop}
\title{Calculating net radiation available to the crop}
\usage{
cal_netRadiationForCrop(pRn, Rn)
}
\arguments{
\item{pRn}{canopy penetration probability for net radiation}

\item{Rn}{the net radiation (W/m2 ground).see cal_hourlyNetRadiation()}
}
\value{
A vector for net radiation available to the crop (W/m2 ground)
}
\description{
Calculating net radiation available to the crop
}
