% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_netRadiationForSoil}
\alias{cal_netRadiationForSoil}
\title{Calculating net radiation available to the soil}
\usage{
cal_netRadiationForSoil(pRn, Rn, G)
}
\arguments{
\item{pRn}{canopy penetration probability for net radiation}

\item{Rn}{the net radiation (W/m2 ground).see cal_hourlyNetRadiation()}

\item{G}{the soil heat flux(W/m2 ground)}
}
\value{
A vector for net radiation available to the soil (W/m2 ground)
}
\description{
Calculating net radiation available to the soil
}
