% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_netRadiationForSystem}
\alias{cal_netRadiationForSystem}
\title{Calculating net radiation available to the system(soil and crop)}
\usage{
cal_netRadiationForSystem(Ac, As)
}
\arguments{
\item{Ac}{net radiation available to the crop (W/m2 ground)}

\item{As}{net radiation available to the soil(W/m2 ground)}
}
\value{
A vector for net radiation available to the system(soil and crop) (W/m2 ground)
}
\description{
Calculating net radiation available to the system(soil and crop)
}
