% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_netSolarRadiation}
\alias{cal_netSolarRadiation}
\title{Calculating net solar (shortware radiation) Rns}
\usage{
cal_netSolarRadiation(alpha, Rs)
}
\arguments{
\item{alpha}{albedo or canopy reflection coefficient, which is 0.23 for the
hypothetical grass reference crop [dimensionless].}

\item{Rs}{the incoming solar radiation [MJ m-2 day-1]. From cal_solarRadiation()}
}
\value{
A vector for net solar or shortwave radiation [MJ m-2 day-1].
}
\description{
The net shortwave radiation resulting from the balance between
   incoming and reflected solar radiation.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
