% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_percolationForExcessWater}
\alias{cal_percolationForExcessWater}
\title{Calculating percolation for excess water}
\usage{
cal_percolationForExcessWater(THETA_i_t0, Pe_i_t1, THETA_sat_i)
}
\arguments{
\item{THETA_i_t0}{the water amount of the day before in soil layer i (mm)}

\item{Pe_i_t1}{the percolation of previous soil layer(mm)}

\item{THETA_sat_i}{soil saturation water amount (mm)}
}
\value{
A value for percolation for excess water (mm)
}
\description{
Calculating percolation for excess water
}
