% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_psychrometriCconstant}
\alias{cal_psychrometriCconstant}
\title{Calculating psychrometric constant}
\usage{
cal_psychrometriCconstant(atmospheric_pressure)
}
\arguments{
\item{atmospheric_pressure}{atmospheric pressure (kPa).}
}
\value{
A vector for Psychrometric constant (kPa/degree Celsius)
}
\description{
Calculating psychrometric constant
}
\examples{
cal_psychrometriCconstant(100.1235)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
