% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_reductionFactorForE}
\alias{cal_reductionFactorForE}
\title{Calculating reduction factor for evaporation}
\usage{
cal_reductionFactorForE(THETA, THETA_sat)
}
\arguments{
\item{THETA}{the water amount of the day before in stop soil layer  (mm)}

\item{THETA_sat}{soil saturation water content in stop oil layer (mm)}
}
\value{
A value for reduction factor for evaporation
}
\description{
Calculating reduction factor for evaporation
}
