% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_reductionFactorForT}
\alias{cal_reductionFactorForT}
\title{calculating reduction factor for transpiration}
\usage{
cal_reductionFactorForT(THETA_v_wp, p, THETA_v_sat, THETA_v)
}
\arguments{
\item{THETA_v_wp}{soil water content at wilting point (m3 m-3)}

\item{p}{a coefficient. 0.5 for C3 and 0.3 for C4 plant}

\item{THETA_v_sat}{soil saturation water content (m3 m-3)}

\item{THETA_v}{the water amount of the day before in root layer(m3 m-3)}
}
\value{
A value for reduction factor for transpiration
}
\description{
calculating reduction factor for transpiration
}
