% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_Rs_from_Na}
\alias{cal_Rs_from_Na}
\title{Calculating Solar radiation from actual duration of sunshine}
\usage{
cal_Rs_from_Na(as = 0.25, bs = 0.5, Na, Latitude, J)
}
\arguments{
\item{as}{regression constant, expressing the fraction of extraterrestrial
radiation reaching the earth on overcast days (n = 0).Default is 0.25.}

\item{bs}{as+bs is fraction of extraterrestrial radiation reaching the earth
on clear days (n = N). Default is 0.50.}

\item{Na}{actual duration of sunshine [hour].}

\item{Latitude}{latitude (angert).}

\item{J}{is the number of the day in the year between 1 (1 January) and 365
or 366 (31 December).}
}
\value{
A vector for solar radiation(MJ m-2 d-1)
}
\description{
Calculating Solar radiation from actual duration of sunshine
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
