% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_saturationVapourPressure}
\alias{cal_saturationVapourPressure}
\title{Calculating saturation vapour pressure}
\usage{
cal_saturationVapourPressure(Tem)
}
\arguments{
\item{Tem}{air temperature (degrees Celsius).}
}
\value{
A vector for saturation vapour pressure at the air temperature T (kPa).
}
\description{
saturation vapour pressure at the air temperature T.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
