% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_sensibleHeatFluxesForCrop}
\alias{cal_sensibleHeatFluxesForCrop}
\title{Calculating sensible heat fluxes for crop}
\usage{
cal_sensibleHeatFluxesForCrop(gamma, Ac, r_c_s, r_c_a, rho_cp, D0, DELTA)
}
\arguments{
\item{gamma}{is the psychometric constant (0.658 mbar K-1)}

\item{Ac}{energy available to the crop (W m-2 ground)}

\item{r_c_s}{the canopy resistance (s m-1)}

\item{r_c_a}{is the bulk boundary layer resistance (s m-1)}

\item{rho_cp}{the volumetric heat capacity for air (1221.09 J m-3 K-1)}

\item{D0}{the vapor pressure deficit at the mean canopy flow}

\item{DELTA}{the slope of the saturated vapor pressure curve (mbar K-1)}
}
\value{
A vector for sensible heat fluxes for crop
}
\description{
Calculating sensible heat fluxes for crop
}
