% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_skySolarRadiation_withas_bs}
\alias{cal_skySolarRadiation_withas_bs}
\title{Calculating clear sky solar radiation with as and bs}
\usage{
cal_skySolarRadiation_withas_bs(as, bs, Ra)
}
\arguments{
\item{as, bs}{as+bs fraction of extraterrestrial radiation reaching the earth
on clear-sky days (n = N).}

\item{Ra}{extraterrestrial radiation [MJ m-2 day-1].
From cal_extraterrestrialRadiation_for_daily()}
}
\value{
A vector for clear-sky solar radiation [MJ m-2 day-1].
}
\description{
The calculation of the clear-sky radiation, Rso, when n = N,
   is required for computing net longwave radiation.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
