% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_soilHeatFlux}
\alias{cal_soilHeatFlux}
\title{Calculating Soil/ground heat flux}
\usage{
cal_soilHeatFlux(pRn, Rn)
}
\arguments{
\item{pRn}{the canopy penetration probility for net radiation}

\item{Rn}{the net radiation (W/m2 ground)}
}
\value{
A vector for the soil heat flux (W/m2 ground)
}
\description{
Calculating Soil/ground heat flux
}
