% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_soilHeatFlux_day}
\alias{cal_soilHeatFlux_day}
\title{Calculating soil heat flux(G) for day/ten-day periods}
\usage{
cal_soilHeatFlux_day()
}
\value{
A value for 0
}
\description{
As the magnitude of the day or ten-day soil heat flux beneath
   the grass reference surface is relatively small, it may be ignored .
}
\references{
FAO Irrigation and drainage paper 56 (P54)
}
