% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_soilHeatFlux_monthly}
\alias{cal_soilHeatFlux_monthly}
\title{Calculating soil heat flux(G) for monthly periods}
\usage{
cal_soilHeatFlux_monthly(T1, T0, Tmonth2 = TRUE)
}
\arguments{
\item{T1}{air temperature at time i [degrees Celsius].}

\item{T0}{air temperature at time i-1 [degrees Celsius].}

\item{Tmonth2}{Is the mean air temperature of next month know?}
}
\value{
A vector for soil heat flux [MJ m-2 day-1]
}
\description{
When assuming a constant soil heat capacity of 2.1 MJ m-3 °C-1
   and an appropriate soil depth, cal_soilHeatFlux_general can be used to
   derive G for monthly periods.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
