% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_soilSurfaceToMeanCanopyFlow}
\alias{cal_soilSurfaceToMeanCanopyFlow}
\title{Calculating soil surface to mean canopy flow}
\usage{
cal_soilSurfaceToMeanCanopyFlow(h, nK, Kh, zs0, z0, d)
}
\arguments{
\item{h}{the plant height(m)}

\item{nK}{the eddy diffusivity extinction coefficient (taken as 2)}

\item{Kh}{eddy diffusivity at the canopy top(m2/s)}

\item{zs0}{is the soil surface roughnesslength (m). Note: for flat, tilled land, zs0 can be taken as 0.004 m.}

\item{z0}{the crop roughness length (m)}

\item{d}{zero plane displacement height (m)}
}
\value{
A vector for aerodynamic resistancessoil surface to mean canopy flow (s m-1)
}
\description{
the resistance between the soil surface and the mean canopy flow (s m-1)
}
