% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_solarDeclination}
\alias{cal_solarDeclination}
\title{Calculating solar declination}
\usage{
cal_solarDeclination(td)
}
\arguments{
\item{td}{is the day of year.}
}
\value{
A vector for solar declination (Radian)
}
\description{
Calculating solar declination
}
\note{
The solar declination actually varies throughout the day too but
   its variation is very small; thus, it is often ignored.
   Negative angles occur when the angle is below the equator plane,
   positive for above the equator.
}
\examples{
cal_solarDeclination(34)
}
\references{
Teh CBS.Introduction to mathematical modeling of crop growth:
           How the equations are derived and assembled into a computer model.
           Brown Walker Press, 2006.
}
