% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_solarDeclination_in_FAO}
\alias{cal_solarDeclination_in_FAO}
\title{Calculating solar declination with FAO56 method}
\usage{
cal_solarDeclination_in_FAO(J)
}
\arguments{
\item{J}{is the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)}
}
\value{
A vector for solar declination(Radian)
}
\description{
Calculating solar declination with FAO56 method
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
