% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_sunsetHourAngle}
\alias{cal_sunsetHourAngle}
\title{Calculating sunset hour angle}
\usage{
cal_sunsetHourAngle(lat, solar_declination)
}
\arguments{
\item{lat}{latitude (Radian), positive for the northern hemisphere and
negative for the southern hemisphere.}

\item{solar_declination}{solar declination(Radina).}
}
\value{
A vector for sunset Hour Angle(Radian)
}
\description{
Calculating sunset hour angle
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
