% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_sunsetTime}
\alias{cal_sunsetTime}
\title{Calculating the local solar time for sunset/sunrise}
\usage{
cal_sunsetTime(solar_declination, latitude)
}
\arguments{
\item{solar_declination}{can be calculated by cal_solardeclination().}

\item{latitude}{is latitude data(Radian).}
}
\value{
A vector for the local solar time for sunset/sunrise
}
\description{
Calculating the local solar time for sunset/sunrise.
}
\details{
Knowing the time of sunrise can calculate the time of sunset.
   sunrise_time=24-sunset_time. Day_length=2*(sunset_time-12).
}
\references{
Teh CBS.Introduction to mathematical modeling of crop growth:
           How the equations are derived and assembled into a computer model.
           Brown Walker Press, 2006.
}
