% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meteorlogy.R
\name{cal_TemMean}
\alias{cal_TemMean}
\title{calculating the mean daily air temperature}
\usage{
cal_TemMean(Tmax, Tmin)
}
\arguments{
\item{Tmax}{the daily maximum.The temperature is given in degree Celsius or Fahrenhei.}

\item{Tmin}{the daily minimum.The temperature is given in degree Celsius,or Fahrenhei.}
}
\value{
A vector for the mean daily air temperature
}
\description{
calculating the mean daily air temperature
}
\details{
It is only employed in the FAO Penman-Monteith equation to calculate
   the slope of the saturation vapor pressure curves and the impact of mean
   air density as the effect of temperature variations on the value of the
   climatic parameter is small in these cases. For standardization, Tmean for
   24-hour periods is defined as the mean of mean of the daily maximum and
   minimum temperatures rather than as the average of hourly temperature
   measurements.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
