% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DualKc.R
\name{cal_TEW_for_DualKc}
\alias{cal_TEW_for_DualKc}
\title{calculating total evaporable water}
\usage{
cal_TEW_for_DualKc(FC, WP, Ze)
}
\arguments{
\item{FC}{Soil water content at field capacity, m3 m-3}

\item{WP}{Soil water content at wilting point, m3 m-3}

\item{Ze}{Depth of the surface soil layer that is subject to drying by way of evaporation, 0.10-0.15 m.}
}
\value{
A value for total evaporable water
}
\description{
maximum depth of water that can be evaporated from the soil
   when the topsoil has been initially completely wetted.
   Estimated TEW for Kr calculation
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
