% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plant.R
\name{cal_WaterStressCoef}
\alias{cal_WaterStressCoef}
\title{Calculating water stress coefficient}
\usage{
cal_WaterStressCoef(Dr, TAW, p)
}
\arguments{
\item{Dr}{root zone depletion(mm).}

\item{TAW}{total available soil water in the root zone(mm).}

\item{p}{fraction of TAW that a crop can extract from the root zone without
suffering water stress.}
}
\value{
A value for water stress coefficient which is a dimensionless transpiration reduction factor
   dependent on available soil water
}
\description{
Calculating water stress coefficient
}
