% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_windSpeed_Canopy}
\alias{cal_windSpeed_Canopy}
\title{Calculating wind speed above and within the canopies.}
\usage{
cal_windSpeed_Canopy(z, h, u_, k = 0.4, d, z0, nu = 2)
}
\arguments{
\item{z}{the height (m)}

\item{h}{the plant height (m)}

\item{u_}{the friction velocity (m/s)}

\item{k}{the von Karman constant (0.4)}

\item{d}{zero plane displacement height (m)}

\item{z0}{the crop roughness length(m)}

\item{nu}{the wind speed extinction coefficient (taken as 2)}
}
\value{
A vector for the wind speed (m/s) at height z(m)
}
\description{
Calculating wind speed above and within the canopies.
}
