% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Shuttleworth_Wallace.R
\name{cal_zeroPlaneHeight}
\alias{cal_zeroPlaneHeight}
\title{Calculating zero plane displacement height}
\usage{
cal_zeroPlaneHeight(h)
}
\arguments{
\item{h}{the plant height (m)}
}
\value{
A vector for zero plane displacement height (m)
}
\description{
Calculating zero plane displacement height
}
