% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{compare_model_plot}
\alias{compare_model_plot}
\title{Show the results of different models}
\usage{
compare_model_plot(model_list, names)
}
\arguments{
\item{model_list}{List. Including output results of different models.}

\item{names}{Vector. Name of models.}
}
\value{
A list for ggplot2 plot
}
\description{
Show the results of different models
}
