% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_angert_to_radian}
\alias{convert_angert_to_radian}
\title{Converting angert to radian}
\usage{
convert_angert_to_radian(anger)
}
\arguments{
\item{anger}{Longitude or dimension in Angle.}
}
\value{
A vector for longitude or dimension in radian.
}
\description{
Converting the unit of angle in longitude and latitude into
   the unit of radian.
}
\examples{
convert_angert_to_radian(98.8)
}
