% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_Date_to_dayofyear}
\alias{convert_Date_to_dayofyear}
\title{Convert date to day of year}
\usage{
convert_Date_to_dayofyear(Date)
}
\arguments{
\item{Date}{is a date format data.}
}
\value{
A vector for the day of year.
}
\description{
Convert date to day of year
}
