% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_degreesCelsius_to_Fahrenheit}
\alias{convert_degreesCelsius_to_Fahrenheit}
\title{Convert degrees Celsius to Fahrenheit}
\usage{
convert_degreesCelsius_to_Fahrenheit(degrees_Celsius)
}
\arguments{
\item{degrees_Celsius}{temperature in degrees Celsius(°C).}
}
\value{
A vector for atemperature in Fahrenheit(°F).
}
\description{
Convert degrees Celsius to Fahrenheit
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
