% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_Fahrenheit_to_degreesCelsius}
\alias{convert_Fahrenheit_to_degreesCelsius}
\title{Convert Fahrenheit to degrees Celsius}
\usage{
convert_Fahrenheit_to_degreesCelsius(Fahrenheit)
}
\arguments{
\item{Fahrenheit}{temperature in Fahrenheit(°F).}
}
\value{
A vector for temperature in degrees Celsius(°C)
}
\description{
Convert Fahrenheit to degrees Celsius
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
