% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_windSpeed_to_2m}
\alias{convert_windSpeed_to_2m}
\title{Convert wind speed to the standard of 2m}
\usage{
convert_windSpeed_to_2m(uz, z)
}
\arguments{
\item{uz}{measured wind speed at z m above ground surface [m s-1].}

\item{z}{height of measurement above ground surface [m].}
}
\value{
A vector for wind speed at 2 m above ground surface [m s-1].
}
\description{
For the calculation of evapotranspiration, wind. Speed measured
   at 2 m above the surface is required. To adjust wind speed data obtained
   from instruments placed at elevations other than the standard height of 2 m,
   a logarithmic wind speed profile may be used for measurements above a short grassed surface.
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
