% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{estimate_goodnessOfFit}
\alias{estimate_goodnessOfFit}
\title{Calculating the goodness-of-fit indicators between measured and simulated values}
\usage{
estimate_goodnessOfFit(Sim, Obs)
}
\arguments{
\item{Sim}{The simualtion value of model.}

\item{Obs}{The observed value.}
}
\value{
A vector for the goodness-of-fit indicators
}
\description{
Calculating the goodness-of-fit indicators between measured and simulated values
}
