% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{estimate_LAI_for_alfalfa}
\alias{estimate_LAI_for_alfalfa}
\title{Estimate LAI for alfalfa}
\usage{
estimate_LAI_for_alfalfa(hc)
}
\arguments{
\item{hc}{is the vegetation height in meter. (in meter)}
}
\value{
A vector for leaf are index of alfalfa
}
\description{
Estimate LAI for alfalfa
}
\references{
Zhao C , Feng Z , Chen G . Soil water balance simulation of alfalfa (Medicago sativa L.)
 in the semiarid Chinese Loess Plateau[J]. Agricultural Water Management, 2004, 69(2):0-114.
}
