% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FIalfalfa}
\alias{FIalfalfa}
\title{A example dataset of alfalfa under flood irrigation}
\format{
A data frame with 161 rows and 22 variables
}
\usage{
FIalfalfa
}
\description{
A example dataset of alfalfa under flood irrigation
}
\keyword{datasets}
