% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_DualKc.R
\name{Kcb_adj_for_DualKc}
\alias{Kcb_adj_for_DualKc}
\title{Adjust the recommended Kc values at the middle and late stages}
\usage{
Kcb_adj_for_DualKc(Kcb_table, u2, RHmin, h)
}
\arguments{
\item{Kcb_table}{Recommended value of KC in FAO 56 at the middle and late stages}

\item{u2}{wind speed at 2 m}

\item{RHmin}{Minimum relative humidity}

\item{h}{Plant height}
}
\value{
A value for adjust Kc at middle and late stages
}
\description{
Adjust the recommended Kc values at the middle and late stages
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M.
   FAO Irrigation and drainage paper No. 56. Rome: Food and Agriculture
   Organization of the United Nations, 1998.
}
