% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{linear interpolation for vector}
\usage{
linear_interpolation(DataVector)
}
\arguments{
\item{DataVector}{data vector.Note that the starting value of vector needs
to be no missing value.}
}
\value{
A interpolated vector
}
\description{
Linear interpolation is performed by using the values on both
   sides of the missing values.
}
