% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SDIalfalfa}
\alias{SDIalfalfa}
\title{A example dataset of alfalfa under subsurface drip irrigation}
\format{
A data frame with 161 rows and 22 variables
}
\usage{
SDIalfalfa
}
\description{
A example dataset of alfalfa under subsurface drip irrigation
}
\keyword{datasets}
