% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simET-package.R
\docType{package}
\name{simET-package}
\alias{simET}
\alias{simET-package}
\title{simET: Tools for Simulation of Evapotranspiration of Field Crops and Soil Water Balance}
\description{
Supports the calculation of meteorological characteristics in evapotranspiration research and reference crop evapotranspiration, and offers three models to simulate crop evapotranspiration and soil water balance in the field, including single crop coefficient and dual crop coefficient, as well as the Shuttleworth-Wallace model. These calculations main refer to Allen et al.(1998, ISBN:92-5-104219-5), Teh (2006, ISBN:1-58-112-998-X), and Liu et al.(2006) <doi:10.1016/j.agwat.2006.01.018>.
}
\author{
\strong{Maintainer}: Minguo Liu \email{liumg15@lzu.edu.cn}

Other contributors:
\itemize{
  \item Huimin Yang \email{huimyang@lzu.edu.cn} [data contributor, funder]
}

}
\keyword{internal}
