% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simET.R
\docType{package}
\name{simET}
\alias{simET}
\title{simET: Evapotranspiration simulation in field}
\description{
The package provides some functions to calculate
some meteorological parameters and ET0, and provides
three models to simulate evapotranspiration,
including single crop coefficient and dual crop
coefficient, Shuttleworth – Wallace model
}
